fix(jkr): widen Revit element ids to long to survive 2027 runtime
new ElementId(intField) bound to ElementId(Int32) — obsolete since the
Revit 2024 64-bit id change, kept in the 2025.* compile ref but removed
from the Revit 2027 runtime. The call compiled, then threw
MissingMethodException on click ("Method not found: ElementId..ctor(Int32)"),
aborting JKR Fix All / Fix / Locate.

Migrate element-id storage int -> long end-to-end so every new ElementId(...)
binds to the runtime-present ElementId(Int64) ctor:

- crash sites: JkrRenameHandler.ApplyRename (the primary rename path),
  JkrFixApplicator preflight, JkrComplianceDashboardPanel.LocateInRevit,
  ComplianceDashboardPanel.SelectElementInRevit
- DTOs: JkrFixAction/FixResult, IssueVm, JkrAuditStore snapshot,
  ComplianceIssueDto/ModelCheckElement/AIRecommendationDto,
  JkrElementData/JkrComplianceCheckV2/JkrFixActionV2
- cascaded signatures/collections: RenameQueue tuple, RenameResult.FailedElementIds
  (HashSet<long>), MakeFixKey, IssueMapper.BlockFix/IsBlocked, ComputeIssueId
- kill the truncation source: JkrBuildingInfoExtractor no longer casts
  elem.Id.Value down to int (also fixes corruption of ids > int.MaxValue)

Existing .jkr_audit.json integer ids deserialize into long unchanged.
Cost subsystem (CostItem/AICostEstimator) left as-is: no ElementId ctor,
cannot hit this crash.

+++ Handlers/JkrRenameHandler.cs
@@ -17,7 +17,7 @@ namespace RevitWebAppSync.Handlers
         /// <summary>
         /// List of (ElementId, newName) pairs to rename.
         /// </summary>
-        public List<(int ElementId, string NewName)> RenameQueue { get; set; } = new List<(int, string)>();
+        public List<(long ElementId, string NewName)> RenameQueue { get; set; } = new List<(long, string)>();
 
         /// <summary>
         /// List of parameter fixes to apply after renames.
@@ -191,7 +191,7 @@ namespace RevitWebAppSync.Handlers
             }
         }
 
-        private static void ApplyRename(Document doc, int elemId, string newName,
+        private static void ApplyRename(Document doc, long elemId, string newName,
                                         RenameResult result, List<string> failReasons)
         {
             try
@@ -331,7 +331,7 @@ namespace RevitWebAppSync.Handlers
         /// element-level granularity. Don't use this when an element can have multiple
         /// fixes (rename + several param fixes) and you need to distinguish which
         /// specific fix failed; use FailedFixKeys instead.</summary>
-        public HashSet<int> FailedElementIds { get; set; } = new HashSet<int>();
+        public HashSet<long> FailedElementIds { get; set; } = new HashSet<long>();
 
         /// <summary>Per-fix failure keys — distinguishes "this rename failed" from
         /// "this Sistem param fix failed" on the same element. Format:
@@ -344,7 +344,7 @@ namespace RevitWebAppSync.Handlers
         public HashSet<string> FailedFixKeys { get; set; } = new HashSet<string>();
 
         /// <summary>Compute a fix key matching the format used by FailedFixKeys.</summary>
-        public static string MakeFixKey(string action, int elementId, string parameterName)
+        public static string MakeFixKey(string action, long elementId, string parameterName)
         {
             if (string.Equals(action, "rename_type", System.StringComparison.OrdinalIgnoreCase))
                 return $"r:{elementId}";
